fix(ui): quokka — suggestions ARE the demo (one typing list) [XS]
User: "you are typing the phrases but still have the static ones. I
was envisioning each one coming in independently in their place and
then that's done."

Before: TypingPrompt demo callout + static .v2-adviser-suggestion
buttons. Phrases appeared twice; terminal mode made them look
identical with the > prefix on both.

After: TypingSuggestions component — single clickable list. Each slot
starts empty (faded), types in sequence, becomes clickable on complete.
No separate demo.

States:
  pending  — opacity 0.55, no cursor, disabled
  active   — full opacity, current text + cursor inside the button
  complete — full opacity, full phrase, clickable

prefers-reduced-motion sets phase 'done' on mount → all clickable.

Deleted TypingPrompt; adjusted AdviserModal.css + terminal/init.css
selectors for the new class names.
