--- src/v2/components/TypingPrompt.css
@@ -1,40 +1,0 @@
-/* Typing-prompt — types each phrase in sequence, builds up a list of
- * completed lines, holds the final state. Used by Quokka's empty-state
- * demo above the static suggestion buttons. */
-
-.v2-typing-prompt {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  width: 100%;
-}
-
-.v2-typing-prompt-line {
-  font: 400 13px/1.5 var(--v2-font-body);
-  color: var(--v2-text);
-  word-break: break-word;
-  overflow-wrap: break-word;
-}
-
-.v2-typing-prompt-line-complete {
-  color: var(--v2-text-meta);
-}
-
-.v2-typing-prompt-cursor {
-  color: var(--v2-accent);
-  font-weight: 700;
-  margin-left: 1px;
-  animation: v2-typing-prompt-blink 1.1s steps(2) infinite;
-}
-
-@keyframes v2-typing-prompt-blink {
-  0%, 49%   { opacity: 1; }
-  50%, 100% { opacity: 0; }
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .v2-typing-prompt-cursor {
-    animation: none;
-    opacity: 0.6;
-  }
-}
